+++ CHANGELOG.md
@@ -34,6 +34,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
  - Follow-mode is now accessible from the Tracker menu
  - Selecting an instrument/waveform will switch to that item in its editor if it is open.
  - Song editor not updating on song change.
+ - Module properties dialog can now be opened via Ctrl+P
 ### Fixed
  - Issue with Instrument Editor not properly updating when changing instruments
  - Hitting enter in the sequence editor line edit opens the Wave editor